feat:收藏餐廳功能、Jwt邏輯 fix:SecurityConfig

+++ foodmap/src/main/java/com/example/foodmap/controller/RestaurantController.java
@@ -21,7 +21,7 @@ public class RestaurantController {
     }
 
     @GetMapping("/restaurants/{id}")
-    public Restaurant getById(@PathVariable int id){
+    public Restaurant getById(@PathVariable Long id){
         return service.getById(id);
     }
